requirements.txt
# Core Framework
streamlit==1.30.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# API Integration
httpx==0.26.0
requests==2.31.0

# Authentication & Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Data Visualization (for reports)
plotly==5.18.0
pandas==2.1.4
